perf: Change libdw libraries to support elfutils>=0.178
Up untill now, static compilation of perf was still trying to dynamically
load the backends for libebl. Starting with version 0.178 of elfutils,
libebl no longer needs to dynamicall load backends. This means that
perf compiled statically with elfutils>=0.178 is trully static.

In elfutils 0.178, libebl is also no longer a standalone library,
it is included as part of libdw, so there is no longer a need for -lebl.
Also - some of the backends required -lpthread, and since they are now
included in the static build, we also need to add -lpthread.

This change does not support older versions of elfutils.

Signed-off-by: Daniel Shaulov <daniel.shaulov@granulate.io>