perf probe: Make perf-probe use symbols when debuginfo is incomplete
When a binary has debug symbols, but they are incomplete,
some files compiled with debug info and some didn't, perf-probe will
look for the debuginfo, and when it fails to find it will refuse to add the probe.
If the debug info is removed entirely perf-probe gracefully degrades to using symbols.
This commit adds the fallback to symbols for cases where dwarf exists but fails

The issue was found with a nodejs binary which ships with debug info only
for the '.c' files but no for '.cc' (c++) files, and trying to probe a c++ function.

Signed-off-by: Daniel Shaulov <daniel.shaulov@granulate.io>