perf probe: Add uint8* as a supported string type
When trying to cast a probe variable to string, the type is checked
to be "(unsigned) char *". In Golang debug info, the strings are
ultimately "uint8 *", so the cast fails.

This commit adds the type to the list of types that can be casted
to string, and also prints the type when it is found to not match
one the types, for easier debugging of future issues.

Signed-off-by: Daniel Shaulov <daniel.shaulov@granulate.io>