perf inject: Fix header after injecting
The cmdline (and possibly other fields) in the header, are being
overwritten by "perf inject". It means that after injecting, running
"perf script --header" will show the cmdline of "perf inject", that
contains limited usefull information, as opposed to the commandline of
"perf record" which explains what was recorded.
This commit fixes it by overwritting the new env with the env from
the file before calling perf_session__write_header.

Signed-off-by: Daniel Shaulov <daniel.shaulov@granulate.io>